refactor duplicated code into "AbstractFavoritesTest"
